requirements.txt
Flask == 2.2.2
Flask-Cors == 3.0.10
Flask-RESTful == 0.3.9
jsonschema ==4.16.0
mysql-connector-python
python-dotenv == 1.0.0
Flask-Testing == 0.8.1
PyJWT == 2.7.0